move RowRanges
